+++ packages/core/spec/Rules/DviToSvg_spec.ts
@@ -1,4 +1,5 @@
 /// <reference path="../../node_modules/@types/jasmine/index.d.ts" />
+/// <reference path="../../node_modules/@types/jasmine-expect/index.d.ts" />
 
 import DviToSvg from '../../src/Rules/DviToSvg'
 import { initializeRule, RuleDefinition } from '../helpers'
@@ -19,7 +20,7 @@ describe('DviToSvg', () => {
         options: { outputFormat: 'svg' }
       })
 
-      expect(await DviToSvg.isApplicable(rule, 'build', 'execute', rule.parameters)).toBe(true)
+      expect(await DviToSvg.isApplicable(rule, 'build', 'execute', rule.parameters)).toBeTrue()
 
       done()
     })
@@ -29,7 +30,7 @@ describe('DviToSvg', () => {
         options: { outputFormat: 'dvi' }
       })
 
-      expect(await DviToSvg.isApplicable(rule, 'build', 'execute', rule.parameters)).toBe(false)
+      expect(await DviToSvg.isApplicable(rule, 'build', 'execute', rule.parameters)).toBeFalse()
 
       done()
     })