+++ packages/core/src/State.ts
@@ -11,8 +11,7 @@ import {
   FileCache,
   KillToken,
   OptionInterfaceMap,
-  Phase,
-  RuleCache
+  Phase
 } from './types'
 
 function getLabel (x: File | Rule): string {
@@ -120,35 +119,6 @@ export default class State extends EventEmitter {
   removeFile (file: File): void {
     this.files.delete(file.filePath)
     this.removeNode(file.filePath)
-  }
-
-  async addCachedRule (cache: RuleCache): Promise<void> {
-    const options: OptionsInterface = this.getJobOptions(cache.jobName)
-    const id: string = this.getRuleId(cache.name, cache.command, cache.phase, cache.jobName, cache.parameters)
-    if (this.rules.has(id)) return
-
-    const RuleClass = this.ruleClasses.find(ruleClass => ruleClass.name === cache.name)
-    if (RuleClass) {
-      const parameters: File[] = []
-      for (const filePath of cache.parameters) {
-        const parameter: File | undefined = await this.getFile(filePath)
-        if (!parameter) break
-        parameters.push(parameter)
-      }
-      const rule: Rule = new RuleClass(this, cache.command, cache.phase, options, parameters)
-      this.addNode(rule.id)
-      await rule.initialize()
-      this.rules.set(rule.id, rule)
-      await rule.getInputs(cache.inputs)
-      const outputs: File[] = await rule.getOutputs(cache.outputs)
-      if (RuleClass.alwaysEvaluate || outputs.length !== cache.outputs.length) {
-        // At least one of the outputs is missing or the rule should always run.
-        rule.addActions()
-      }
-      for (const input of rule.inputs) {
-        await rule.addFileActions(input)
-      }
-    }
   }
 
   getRuleId (name: string, command: Command, phase: Phase, jobName: string | null = null, parameters: string[] = []): string {